+++ R/sjPlotFA.R
@@ -35,7 +35,7 @@
 #'
 #' @inheritParams sjp.pca
 #' @inheritParams sjt.pca
-#' @inheritParams sjp.grpfrq
+#' @inheritParams plot_grpfrq
 #'
 #' @note This method for factor analysis relies on the functions
 #'       \code{\link[psych]{fa}} and \code{\link[psych]{fa.parallel}}