+++ R/sjPlotxtab.R
@@ -1,5 +1,5 @@
 #' @title Plot contingency tables
-#' @name sjp.xtab
+#' @name plot_xtab
 #'
 #' @description Plot proportional crosstables (contingency tables) of two variables as ggplot diagram.
 #'
@@ -20,7 +20,7 @@
 #'          with the sum within each category and it's percentages will be added
 #'          to each category.
 #'
-#' @inheritParams sjp.grpfrq
+#' @inheritParams plot_grpfrq
 #'
 #' @return A ggplot-object.
 #'
@@ -31,15 +31,15 @@
 #' x <- sample(1:3, 100, replace = TRUE)
 #'
 #' # plot "cross tablulation" of x and grp
-#' sjp.xtab(x, grp)
+#' plot_xtab(x, grp)
 #'
 #' # plot "cross tablulation" of x and y, including labels
-#' sjp.xtab(x, grp, axis.labels = c("low", "mid", "high"),
+#' plot_xtab(x, grp, axis.labels = c("low", "mid", "high"),
 #'          legend.labels = c("Grp 1", "Grp 2", "Grp 3", "Grp 4"))
 #'
 #' # plot "cross tablulation" of x and grp
 #' # as stacked proportional bars
-#' sjp.xtab(x, grp, margin = "row", bar.pos = "stack",
+#' plot_xtab(x, grp, margin = "row", bar.pos = "stack",
 #'          show.summary = TRUE, coord.flip = TRUE)
 #'
 #' # example with vertical labels
@@ -47,7 +47,7 @@
 #' library(sjlabelled)
 #' data(efc)
 #' set_theme(geom.label.angle = 90)
-#' sjp.xtab(efc$e42dep, efc$e16sex, vjust = "center", hjust = "bottom")
+#' plot_xtab(efc$e42dep, efc$e16sex, vjust = "center", hjust = "bottom")
 #'
 #' # grouped bars with EUROFAMCARE sample dataset
 #' # dataset was importet from an SPSS-file,
@@ -56,11 +56,11 @@
 #' efc.val <- get_labels(efc)
 #' efc.var <- get_label(efc)
 #'
-#' sjp.xtab(efc$e42dep, efc$e16sex, title = efc.var['e42dep'],
+#' plot_xtab(efc$e42dep, efc$e16sex, title = efc.var['e42dep'],
 #'          axis.labels = efc.val[['e42dep']], legend.title = efc.var['e16sex'],
 #'          legend.labels = efc.val[['e16sex']])
 #'
-#' sjp.xtab(efc$e16sex, efc$e42dep, title = efc.var['e16sex'],
+#' plot_xtab(efc$e16sex, efc$e42dep, title = efc.var['e16sex'],
 #'          axis.labels = efc.val[['e16sex']], legend.title = efc.var['e42dep'],
 #'          legend.labels = efc.val[['e42dep']])
 #'
@@ -69,9 +69,9 @@
 #' # so no need to specify labels. For
 #' # title-auto-detection, use NULL
 #' # -------------------------------
-#' sjp.xtab(efc$e16sex, efc$e42dep, title = NULL)
+#' plot_xtab(efc$e16sex, efc$e42dep, title = NULL)
 #'
-#' sjp.xtab(efc$e16sex, efc$e42dep, margin = "row",
+#' plot_xtab(efc$e16sex, efc$e42dep, margin = "row",
 #'          bar.pos = "stack", coord.flip = TRUE)
 #'
 #' @import ggplot2
@@ -80,7 +80,7 @@
 #' @importFrom scales percent
 #' @importFrom stats na.omit
 #' @export
-sjp.xtab <- function(x,
+plot_xtab <- function(x,
                      grp,
                      type = c("bar", "line"),
                      margin = c("col", "cell", "row"),
@@ -467,3 +467,8 @@ sjp.xtab <- function(x,
     legend.labels
   )
 }
+
+
+#' @rdname plot_xtab
+#' @export
+sjp.xtab <- plot_xtab