+++ R/sjTabFA.R
@@ -16,7 +16,7 @@
 #' @inheritParams tab_model
 #' @inheritParams tab_df
 #' @inheritParams sjt.xtab
-#' @inheritParams sjp.grpfrq
+#' @inheritParams plot_grpfrq
 #' @inheritParams sjt.corr
 #'
 #' @return Invisibly returns