+++ R/plot_model.R
@@ -27,6 +27,8 @@
 #'     \item{\code{type = "eff"}}{Similar to \code{type = "pred"}, however,
 #'     discrete predictors are held constant at their proportions (not reference
 #'     level). See \code{\link[ggeffects]{ggeffect}} for details.}
+#'     \item{\code{type = "emm"}}{Similar to \code{type = "eff"}, see
+#'     \code{\link[ggeffects]{ggemmeans}} for details.}
 #'     \item{\code{type = "int"}}{Marginal effects of interaction terms in
 #'     \code{model}.}
 #'   }